Alteração do Routh para [Auth/View - Autenticação e Vizualização] dentro da aplicação
