+++ src/app/templates/editar.html
@@ -17,7 +17,7 @@
 <div class="page-wrapper" id="popup">
     <div class="popup-bg" id="popupBg"></div>
     <div class="popup popup-editar">
-        <form action="{{ url_for('main.edit') }}" method="POST">
+        <form action="{{ url_for('view.edit') }}" method="POST">
             <div class="content-popup-editar">
                 <h2>Novo Informativo</h2>
                 <p>Enviar para:</p>
