+++ src/app/templates/login.html
@@ -5,7 +5,7 @@
 {%block loginTitle %}Bem vindo{% endblock %}
 {%block loginSub %}Entre com seu email da fatec{% endblock %}
 
-{% block form %}{{ url_for('main.login') }}{% endblock %}
+{% block form %}{{ url_for('auth.login') }}{% endblock %}
 
 {% block login %}
