+++ src/app/templates/registrar.html
@@ -5,7 +5,7 @@
 {%block loginTitle %}Registrar{% endblock %}
 {%block loginSub %}Crie uma conta com seu email fatec{% endblock %}
 
-{% block form %}{{ url_for('main.register') }} {% endblock %}
+{% block form %}{{ url_for('auth.register') }} {% endblock %}
 
 {% block login %}
 <div class="login-input registrar-input">
